Support indexing and 'length' field access in incomplete array types
