+++ src/Core/ScriptLang/Ast/Types/IncompleteArrayType.cs
@@ -1,17 +1,16 @@
 ﻿namespace ScTools.ScriptLang.Ast.Types
 {
+    using ScTools.ScriptLang.Ast.Errors;
+
     /// <summary>
     /// Represents an array without specific size.
     /// Only allowed as function parameter wrapped in <see cref="RefType"/>.
     /// </summary>
-    public sealed class IncompleteArrayType : BaseType, IArrayType
+    public sealed class IncompleteArrayType : BaseArrayType
     {
-        public IType ItemType { get; set; }
-
         public override int SizeOf => 0;
 
-        public IncompleteArrayType(SourceRange source, IType itemType) : base(source)
-            => ItemType = itemType;
+        public IncompleteArrayType(SourceRange source, IType itemType) : base(source, itemType) { }
 
         public override TReturn Accept<TReturn, TParam>(IVisitor<TReturn, TParam> visitor, TParam param)
             => visitor.Visit(this, param);
@@ -22,5 +21,7 @@
         // incomplete array types can reference arrays of any size if their item types are equivalent
         public override bool CanBindRefTo(IType other)
             => other is IArrayType otherArray && ItemType.Equivalent(otherArray.ItemType);
+
+        public override bool CanAssign(IType rhs) => rhs is ErrorType;
     }
 }
